Show an error when the game files are locked when trying to open a mod.
It only checks the game files because Godot can't actually check if the files are locked or not. Godot returns that the file doesn't exist if they are locked and most of the mod files may not actually exist so to avoid false positives I've only add an error if the file is a game file.